alx_travel_app/requirements.txt
Django>=4.2,<5.0
celery>=5.5,<5.6
django-cors-headers>=4.3,<5.0
django-environ>=0.12,<0.13
djangorestframework>=3.14,<3.15
drf-yasg>=1.21,<1.22
mysqlclient>=2.2,<3.0
requests
